+++ tests/test_cli.py
@@ -266,6 +266,76 @@ def test_parse_inputs_from_file():
 
 
 @pytest.mark.cli
+def test_parse_inputs_from_file_relative_paths():
+    """
+    Validate relative file resolutions from execution input file.
+
+    When a 'inputs' file is provided for execution, and contains relative local paths, they should consider the relative
+    location of the file prior to the :term:`CLI` ``PWD``. This common pattern allows the inputs file definition and all
+    its contained references to remain consistent when moved together between locations without hard-coded paths.
+
+    This handling is performed during the '_upload_files' step since only this operation needs to resolve a local path.
+    Therefore, we mock the actual 'upload' operation to avoid the request and validate the path resolution worked.
+    If it didn't work, an error would have been raised before (file not found), or it would not be mapped to a vault
+    in the case of non-local (file://) reference.
+    """
+    local_inputs = []  # to be filled by mock on vault uploads (therefore local file resolved)
+
+    def mock_describe(*_, **__):
+        return OperationResult(False, code=500)
+
+    def mock_upload(href, *_, **__):
+        local_inputs.append(href)
+        f_id = str(uuid.uuid4())
+        body = {"file_href": f"vault://{f_id}", "file_id": f_id, "access_token": f_id}
+        return OperationResult(True, code=200, body=body)
+
+    cwd = os.getcwd()
+    try:
+        with ExitStack() as stack:
+            # use 'describe' to early-stop the operation, since it is the step right after parsing and upload of inputs
+            stack.enter_context(mock.patch("weaver.cli.WeaverClient.describe", side_effect=mock_describe))
+            stack.enter_context(mock.patch("weaver.cli.WeaverClient.upload", side_effect=mock_upload))
+            tmp_dir1 = stack.enter_context(tempfile.TemporaryDirectory())
+            tmp_dir2 = stack.enter_context(tempfile.TemporaryDirectory())
+            os.chdir(tmp_dir2)  # ensure that PWD-based relative path resolution still works after job-file path
+
+            path1 = "./local-file.txt"
+            file1 = os.path.abspath(os.path.join(tmp_dir1, path1))
+            path2 = "./nested/local-file.txt"
+            file2 = os.path.abspath(os.path.join(tmp_dir1, path2))
+            path3 = "./other-file.txt"
+            file3 = os.path.abspath(os.path.join(tmp_dir1, path3))
+            path4 = "./other-dir-file.txt"
+            file4 = os.path.abspath(os.path.join(tmp_dir2, path4))
+            for file in [file1, file2, file3, file4]:
+                os.makedirs(os.path.dirname(file), exist_ok=True)
+                with open(file, "w") as f:
+                    f.write("test")
+
+            cwl_inputs = {
+                "input1": "data",  # not affected since not a file reference
+                "input2": {"class": "File", "path": "https://fake.domain.com/some-file.txt"},  # ignore not local file
+                "input3": {"class": "File", "path": path1},
+                "input4": [
+                    {"class": "File", "path": path2},
+                    {"class": "File", "path": f"file://{file3}"},  # ensure absolute path still resolves
+                ],
+                "input5": {"class": "File", "path": path4},  # relative to PWD
+            }
+            cwl_inputs_path = os.path.join(tmp_dir1, "inputs.json")
+            with open(cwl_inputs_path, "w") as cwl_inputs_file:
+                json.dump(cwl_inputs, cwl_inputs_file)
+
+            result = WeaverClient().execute("fake_process", inputs=cwl_inputs_path, url="https://fake.domain.com")
+    finally:
+        os.chdir(cwd)
+    assert result.code == 500, "expected early-stop of execution operation"
+    assert len(local_inputs) == 4
+    assert local_inputs == [file1, file2, file3, file4], "expected local paths to be resolved"
+
+
+@pytest.mark.cli
 @pytest.mark.parametrize(
     ["data_inputs", "expect_inputs"],
     [